+++ tests/testthat/test-response.r
@@ -36,10 +36,10 @@ test_that("status converted to errors", {
   s404 <- GET("http://httpbin.org/status/404")
   s500 <- GET("http://httpbin.org/status/500")
 
-  expect_equal(stop_for_status(s200), TRUE)
-  expect_error(stop_for_status(s300), c("redirection.*\\(300\\)"))
-  expect_error(stop_for_status(s404), c("client.*\\(404\\)"))
-  expect_error(stop_for_status(s500), c("server.*\\(500\\)"))
+  expect_equal(stop_for_status(s200), s200)
+  expect_error(stop_for_status(s300), "redirection.*\\(300\\)")
+  expect_error(stop_for_status(s404), "client.*\\(404\\)")
+  expect_error(stop_for_status(s500), "server.*\\(500\\)")
 })
 
 test_that("headers returned as expected", {